pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "omctl"
version = "0.1.0"
description = "kubectl-style CLI for OpenMetadata — metadata management at the speed of thought"
requires-python = ">=3.9"
keywords = ["openmetadata", "cli", "metadata", "data-catalog", "data-governance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Topic :: Database",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "typer[all]>=0.12.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.scripts]
omctl = "omctl.main:app"

[tool.hatch.build.targets.wheel]
packages = ["omctl"]
